move customize file from init.el to customize.el
Because init.el is now symlinked from the nix store, it's
read-only. This is a problem for Customize because, on invocation, it
tries to write to the custom-file with some defaults. I never use
Customize, but some things (like looking up packages with package.el)
will trigger it behind-the-scenes and then throw errors about init.el
being read-only. Thus setting a specific custom-file will work around
this annoyance.