Minor cleanup
I won't be adding other OSes to my VM, so no need to probe for
them. I really should just ditch GRUB altogether...

Not sure if the pulse snipped is really important and I haven't
actually tried piping audio through the container yet so leaving the
snippet here in case it is needed.