A few more minor LXC tweaks
The kitty terminal requires MIT-SHM which doesn't behave without
having shared IPC which I can't get LXC to place nicely with.

Bootstrap script now passes down the password from the host to the
container. Not applicable for scenarios where the container user is
different from the user on the host bootstrapping the container (that
probably breaks more assumptions I'm making in this script anyway
though).