Launchable xserver from within lxc container
To make this work, it has a lot of nuances and assumptions:

- The host xserver is running on :0

- The host keyboard and mouse device are hardcoded (udev doesn't run
inside LXC)

- The bootstrap lxc.sh script enumerates through a subset of /dev/*
devices and mounts them into the container

- The guest container does not run a full display manager and only
uses startx to launch the xserver

- Lots of things probably don't work (audio, camera, glxinfo are all
untested)

Right now, this container is bootstrapped from NixOS, launched on
NixOS and configured to pass through NixOS-controlled resources. This
will likely need to be made more portable if I want to run it on
Ubuntu, either by adapting the bootstrap script to be runnable on
non-NixOS hosts or do the bootstrapping in a Docker image or
similar. The same goes for the X11 profile generation.