Switch to only-internal if both displays are enabled
When booting the laptop with an external display plugged in, both
displays are enabled by default. The goal of the swap-display script
is to keep only one display enabled and swap between them, so instead
of bailing out (and requiring me to manually turn a display off) just
fall back to the internal display.